fix compiler warnings

+++ src/euklid/vector/vector.cpp
@@ -2,7 +2,7 @@
 
 template<unsigned int dimensions, typename T>
 Vector<dimensions, T>::Vector(const T& in) {
-    for(int i=0; i<dimensions; i++){
+    for(unsigned int i=0; i<dimensions; i++){
         this->set_item(i, in.get_item(i));
     }
 }
@@ -10,7 +10,7 @@ Vector<dimensions, T>::Vector(const T& in) {
 
 template<unsigned int dimensions, typename T>
 Vector<dimensions, T>::Vector(const Vector<dimensions, T>& in) {
-    for(int i=0; i<dimensions; i++){
+    for(unsigned int i=0; i<dimensions; i++){
         this->set_item(i, in.get_item(i));
     }
 }
@@ -18,14 +18,14 @@ Vector<dimensions, T>::Vector(const Vector<dimensions, T>& in) {
 
 template<unsigned int dimensions, typename T>
 Vector<dimensions, T>::Vector() {
-    for(int i=0; i<dimensions; i++){
+    for(unsigned int i=0; i<dimensions; i++){
         this->set_item(i, 0);
     }
 }
 
 template<unsigned int dimensions, typename T>
 Vector<dimensions, T>::Vector(double value) {
-    for(int i=0; i<dimensions; i++){
+    for(unsigned int i=0; i<dimensions; i++){
         this->set_item(i, value);
     }
 }
@@ -35,7 +35,7 @@ template<unsigned int dimensions, typename T>
 double Vector<dimensions, T>::dot(const Vector<dimensions, T>& v2) {
     double result = 0;
 
-    for (int i=0; i<dimensions; i++) {
+    for (unsigned int i=0; i<dimensions; i++) {
         result += this->get_item(i) * v2.get_item(i);
     }
 
@@ -46,7 +46,7 @@ template<unsigned int dimensions, typename T>
 double Vector<dimensions, T>::dot(const T& v2) {
     double result = 0;
 
-    for (int i=0; i<dimensions; i++) {
+    for (unsigned int i=0; i<dimensions; i++) {
         result += this->get_item(i) * v2.get_item(i);
     }
 
@@ -61,7 +61,7 @@ double Vector<dimensions, T>::length(){
 
 template<unsigned int dimensions, typename T>
 double Vector<dimensions, T>::get_item(int n) const {
-    if (n < 0 || n >= dimensions) {
+    if (n < 0 || n >= (int)dimensions) {
         throw "Invalid dimension";
     }
 
@@ -71,7 +71,7 @@ double Vector<dimensions, T>::get_item(int n) const {
 
 template<unsigned int dimensions, typename T>
 void Vector<dimensions, T>::set_item(int n, double value) {
-    if (n < 0 || n >= dimensions) {
+    if (n < 0 || n >= (int)dimensions) {
         throw "Invalid dimension";
     }
 
@@ -80,7 +80,7 @@ void Vector<dimensions, T>::set_item(int n, double value) {
 
 template<unsigned int dimensions, typename T>
 double& Vector<dimensions, T>::operator[](int n) {
-    if (n < 0 || n >= dimensions) {
+    if (n < 0 || n >= (int)dimensions) {
         throw "Invalid dimension";
     }
 
@@ -97,7 +97,7 @@ template<unsigned int dimensions, typename T>
 T Vector<dimensions, T>::operator+(const T& v2) const {
     T result;
 
-    for (int i=0; i<dimensions; i++){
+    for (unsigned int i=0; i<dimensions; i++){
         result.set_item(i, this->get_item(i) + v2.get_item(i));
     }
 
@@ -108,7 +108,7 @@ template<unsigned int dimensions, typename T>
 T Vector<dimensions, T>::operator-(const T& v2) const {
     T result;
 
-    for(int i=0; i<dimensions; i++) {
+    for(unsigned int i=0; i<dimensions; i++) {
         result.set_item(i, this->get_item(i) - v2.get_item(i));
     }
 
@@ -119,7 +119,7 @@ template<unsigned int dimensions, typename T>
 T Vector<dimensions, T>::operator*(const T& v2) const {
     T result;
 
-    for (int i=0; i<dimensions; i++){
+    for (unsigned int i=0; i<dimensions; i++){
         result.set_item(i, this->get_item(i) * v2.get_item(i));
     }
 
@@ -131,7 +131,7 @@ template<unsigned int dimensions, typename T>
 T Vector<dimensions, T>::operator* (const double& factor) const {
     T result;
 
-    for (int i=0; i<dimensions; i++) {
+    for (unsigned int i=0; i<dimensions; i++) {
         result.set_item(i, this->get_item(i)*factor);
     }
 
@@ -152,7 +152,7 @@ template<unsigned int dimensions, typename T>
 void Vector<dimensions, T>::normalize() {
     double len = this->length();
 
-    for (int i=0; i<dimensions; i++) {
+    for (unsigned int i=0; i<dimensions; i++) {
         this->set_item(i, this->get_item(i)/len);
     }
 }
@@ -187,12 +187,12 @@ Vector2D::Vector2D() {
 
 
 Vector2D::Vector2D(const Vector<2, Vector2D>& in) {
-    for(int i=0; i<Vector2D::dimension; i++){
+    for(unsigned int i=0; i<Vector2D::dimension; i++){
         this->set_item(i, in.get_item(i));
     }
 }
 Vector3D::Vector3D(const Vector<3, Vector3D>& in) {
-    for(int i=0; i<Vector3D::dimension; i++){
+    for(unsigned int i=0; i<Vector3D::dimension; i++){
         this->set_item(i, in.get_item(i));
     }
 }
